Move public logo to dedicated componntn
